requirements.txt
langchain_openai 
langchain_core
langchain
openai
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia                                   
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
huggingface_hub
python-dotenv
